+++ xarray/test/test_dask.py
@@ -258,7 +258,7 @@ class TestDataArrayAndDataset(DaskTestCase):
         u = self.eager_array
         v = self.lazy_array
 
-        expected = u.assign_coords(x=u['x'].astype(str))
+        expected = u.assign_coords(x=u['x'])
         self.assertLazyAndIdentical(expected, v.to_dataset('x').to_array('x'))
 
     def test_ufuncs(self):
@@ -300,3 +300,8 @@ class TestDataArrayAndDataset(DaskTestCase):
                              dims=['w', 'z'])
         assert stacked.data.chunks == expected.data.chunks
         self.assertLazyAndIdentical(expected, stacked)
+
+    def test_dot(self):
+        eager = self.eager_array.dot(self.eager_array[0])
+        lazy = self.lazy_array.dot(self.lazy_array[0])
+        self.assertLazyAndAllClose(eager, lazy)